+++ ping.h
@@ -18,12 +18,12 @@ struct ping_option {
 
 struct ping_resp {
     uint32_t total_count;
-    uint32_t resp_time;
+    float resp_time;
     uint32_t seqno;
     uint32_t timeout_count;
     uint32_t bytes;
     uint32_t total_bytes;
-    uint32_t total_time;
+    float total_time;
     int8_t  ping_err;
 };
